Fix int64 build

+++ src/Turbulence/ThHyd/Modeles_Turbulence/RANS/Hydr/Modele_turbulence_hyd_K_Omega.cpp
@@ -402,9 +402,9 @@ void Modele_turbulence_hyd_K_Omega::mettre_a_jour(double temps)
   if (equation().probleme().is_dilatable())
     diviser_par_rho_si_dilatable(ch_K_Omega.valeurs(), mil);
   imprimer_evolution_komega(ch_K_Omega, eqn_transp_K_Omega().schema_temps(), 1);
-  Debog::verifier("Modele_turbulence_hyd_K_Omega::mettre_a_jour loiparoi before", 0);
+  Debog::verifier("Modele_turbulence_hyd_K_Omega::mettre_a_jour loiparoi before", (int)0);
   loipar.calculer_hyd(ch_K_Omega);
-  Debog::verifier("Modele_turbulence_hyd_K_Omega::mettre_a_jour loiparoi after", 1);
+  Debog::verifier("Modele_turbulence_hyd_K_Omega::mettre_a_jour loiparoi after", (int)1);
   eqn_transp_K_Omega().controler_K_Omega();
 
   calculer_viscosite_turbulente(ch_K_Omega.temps());
